Change name of global `~/.gitignore` to `~/.gitignore_global`
Stow, by default, will not copy common VCS metadata files like
`.gitignore`. We can change this behavior by specifying an empty
`~/.stow-global-ignore`, but that would require running `stow` twice to
bootstrap a new system: once to put that file in place, and again for it
to be read, allowing our global gitignore to be copied. Since
`.gitconfig` is copied normally, configuring a different
`core.excludesfile` instead to sidestep issue.